Refactor UIs and Views to their own files
