+++ test/spring-boot/src/test/java/com/example/VaadinSpringBootURIFragmentNavigatorIT.java
@@ -13,6 +13,9 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
+import com.example.view.DefaultView;
+import com.example.view.UIScopedView;
+import com.example.view.ViewScopedView;
 import com.vaadin.testbench.ScreenshotOnFailureRule;
 import com.vaadin.testbench.TestBench;
 import com.vaadin.testbench.TestBenchTestCase;
